♻️ create_env
